// code/src/types/curriculum.ts
export interface ICurriculum {
  id?: string;
  user_id: string;
  name: string;
  layout: string;
  profile_ids?: string[];
  work_ids?: string[];
  formation_ids?: string[];
  skill_ids?: string[];
  language_ids?: string[];
  feedback_ids?: string[];
  address_ids?: string[];
  link_ids?: string[];
  created_at?: string;
  updated_at?: string;
}